requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
beautifulsoup4>=4.9.3
aiohttp>=3.7.4
asyncio>=3.4.3
urllib3>=1.26.5 